project/requirements.txt
pydantic
fastapi
uvicorn
jwt
pyjwt
requests
pytest
httpx
python-dotenv